feat: 로컬 캐싱 구현 및 호출 횟수 줄이기
